refactor #36: SetLocation, LocationButton 컴포넌트에서 useMutation 커스텀훅으로 분리
- SetLocation에 있던 deleteMutation을 useDeleteLocation 훅으로 분리
- LocationButton에 있던 selectMutation을 useSelectLocation 훅으로 분리
